Extract generic offset slider
