+++ OTAnalytics/plugin_ui/nicegui_gui/pages/add_track_form/container.py
@@ -7,7 +7,7 @@ from OTAnalytics.application.resources.resource_manager import (
 from OTAnalytics.plugin_ui.nicegui_gui.pages.add_track_form.add_tracks_form import (
     AddTracksForm,
 )
-from OTAnalytics.plugin_ui.nicegui_gui.pages.add_track_form.offset_slider_form import (
+from OTAnalytics.plugin_ui.nicegui_gui.pages.add_track_form.visualization_offset_slider_form import (  # noqa
     VisualizationOffSetSliderForm,
 )
 from OTAnalytics.plugin_ui.nicegui_gui.pages.add_video_form.container import (
